requirements.txt
fastapi
google - genai
openai
Pillow
playwright
pydantic
